refactor(typecheck): fold tier validator into report-public-contract.mjs --check
One script for the public-contract tier metadata, two modes:
- default: human-readable report (read-only)
- --check: enforcing gate, fails on any invariant violation

Keeps the same pure validator (now exported from report-public-contract.mjs)
and the same 17 unit tests (renamed to scripts/report-public-contract.test.mjs).
Wrapper updated to call the consolidated script.
