+++ packages/superdoc/scripts/README.md
@@ -105,19 +105,28 @@ it stopped running.
 | `report-declaration-reachability.cjs` | postbuild | Instrumentation (not a gate): per-bucket reachability ratio of emitted declarations. | Loses visibility into unreachable emit (the SD-2952 trim target). |
 | `check-jsdoc.cjs` | CI step | Per-file checkJs gate for files in a hand-curated `CHECKED_FILES` allowlist. Currently 6 files. **Note**: `SuperDoc.js` now has `// @ts-check` but is gated by `check:types`, not this script. The 6-file list is a historical ratchet from before the broader enablement; consolidating with `check:types` is tracked separately. | A targeted regression on one of the 6 ratcheted files ships silently. |
 
-The repo also has a top-level tier-discipline gate at
-`scripts/check-public-contract-tiers.mjs`. It runs as stage 1 of
-`check:public:superdoc` (cheap, fast-fail) and enforces the
-`publicContract` taxonomy in `type-surface.config.cjs`:
+The repo also has a top-level tier-discipline gate. One script,
+`scripts/report-public-contract.mjs`, with two modes:
+
+- default (read-only report) - what `pnpm report:public:superdoc` runs.
+  Prints the tiers + a validator status block. Exit 0 always.
+- `--check` (gate) - runs as stage 2 of `check:public:superdoc` after
+  the validator's unit tests. Fails the build on any invariant
+  violation.
+
+Both modes share the pure `validatePublicContract` exported from the
+same file (unit-tested in `scripts/report-public-contract.test.mjs`).
+
+Invariants enforced in `--check` mode against the `publicContract`
+taxonomy in `type-surface.config.cjs`:
 
 - every `package.json#exports` subpath has a tier entry
 - every tier entry exists in `package.json#exports`
 - no subpath appears in more than one tier
+- each entry's `tier` field matches its bucket
 - `supported` subpaths route through `dist/superdoc/src/public/**` (excluding the `legacy/` subtree)
 - `legacy` subpaths route through `dist/superdoc/src/public/legacy/**`
 - `legacyRaw` is restricted to the explicitly accepted set (currently only `./super-editor`)
-
-`report:public:superdoc` is the read-only sibling for human inspection.
 
 ---
